+++ BinarySearchTree/tests.py
@@ -1,17 +1,25 @@
 import unittest
-from binary_tree import BinaryTree
+from binary_search_tree import BinarySearchTree
 
 
 class TestBinaryTree(unittest.TestCase):
 
     def test_init(self):
-        btree = BinaryTree(0)
+        btree = BinarySearchTree()
+        self.assertEqual(btree.data, None)
+        self.assertEqual(btree.left, None)
+        self.assertEqual(btree.right, None)
+        btree = BinarySearchTree(0)
         self.assertEqual(btree.data, 0)
         self.assertEqual(btree.left, None)
         self.assertEqual(btree.right, None)
+        btree = BinarySearchTree([2, 1, 3])
+        self.assertEqual(btree.data, 2)
+        self.assertEqual(btree.left.data, 1)
+        self.assertEqual(btree.right.data, 3)
 
     def test_insert(self):
-        btree = BinaryTree(2)
+        btree = BinarySearchTree(2)
         btree.insert(1)
         self.assertEqual(btree.left.data, 1)
         self.assertEqual(btree.right, None)
@@ -26,7 +34,7 @@ class TestBinaryTree(unittest.TestCase):
         self.assertEqual(btree.right.right.data, 4)
 
     def test_contains(self):
-        btree = BinaryTree(2)
+        btree = BinarySearchTree(2)
         btree.insert(1)
         btree.insert(3)
         btree.insert(0)
@@ -40,7 +48,7 @@ class TestBinaryTree(unittest.TestCase):
         self.assertFalse(btree.contains(-1))
 
     def test_sum(self):
-        btree = BinaryTree(2)
+        btree = BinarySearchTree(2)
         self.assertEqual(btree.sum(), 2)
         btree.insert(1)
         btree.insert(3)
@@ -49,39 +57,39 @@ class TestBinaryTree(unittest.TestCase):
         self.assertEqual(btree.sum(), 10)
 
     def test_is_unival(self):
-        btree = BinaryTree(0)
+        btree = BinarySearchTree(0)
         self.assertTrue(btree.is_unival())
 
-        btree = BinaryTree(5)
+        btree = BinarySearchTree(5)
         btree.insert(5)
         btree.insert(5)
         btree.insert(5)
         self.assertTrue(btree.is_unival())
 
-        btree = BinaryTree(5)
+        btree = BinarySearchTree(5)
         btree.insert(5)
         btree.insert(5)
         btree.insert(1)
         self.assertFalse(btree.is_unival())
 
-        btree = BinaryTree(1)
+        btree = BinarySearchTree(1)
         btree.insert(5)
         btree.insert(5)
         self.assertFalse(btree.is_unival())
 
     def test_is_node_unival(self):
-        btree = BinaryTree(1)
+        btree = BinarySearchTree(1)
         self.assertTrue(btree.is_node_unival())
         btree.insert(1)
         self.assertTrue(btree.is_node_unival())
         btree.insert(1)
         self.assertTrue(btree.is_node_unival())
-        btree = BinaryTree(1)
+        btree = BinarySearchTree(1)
         btree.insert(2)
         self.assertFalse(btree.is_node_unival())
 
     def test_count_univals(self):
-        btree = BinaryTree(5)
+        btree = BinarySearchTree(5)
         btree.insert(5)
         self.assertEqual(btree.count_univals(), 2)
         btree.insert(3)
@@ -91,7 +99,7 @@ class TestBinaryTree(unittest.TestCase):
         self.assertEqual(btree.count_univals(), 5)
 
     def test_max_depth(self):
-        btree = BinaryTree(2)
+        btree = BinarySearchTree(2)
         self.assertEqual(btree.max_depth(), 1)
         btree.insert(3)
         btree.insert(1)
@@ -101,7 +109,7 @@ class TestBinaryTree(unittest.TestCase):
         self.assertEqual(btree.max_depth(), 4)
 
     def test_min_depth(self):
-        btree = BinaryTree(2)
+        btree = BinarySearchTree(2)
         self.assertEqual(btree.min_depth(), 1)
         btree.insert(3)
         btree.insert(1)
@@ -113,7 +121,7 @@ class TestBinaryTree(unittest.TestCase):
         self.assertEqual(btree.min_depth(), 3)
 
     def test_has_path_sum(self):
-        btree = BinaryTree(2)
+        btree = BinarySearchTree(2)
         self.assertTrue(btree.has_path_sum(2))
         self.assertFalse(btree.has_path_sum(5))
         btree.insert(5)
@@ -128,7 +136,7 @@ class TestBinaryTree(unittest.TestCase):
         self.assertFalse(btree.has_path_sum(5))
 
     def test_convert_to_array(self):
-        btree = BinaryTree(2)
+        btree = BinarySearchTree(2)
         self.assertEqual(btree.convert_to_array(), [2])
         btree.insert(1)
         btree.insert(3)
@@ -139,7 +147,7 @@ class TestBinaryTree(unittest.TestCase):
         self.assertEqual(btree.convert_to_array(), [0, 1, 2, 3, 4, 5])
 
     def test_balance(self):
-        btree = BinaryTree(3)
+        btree = BinarySearchTree(3)
         btree.insert(2)
         btree.insert(1)
         btree.insert(0)
@@ -154,7 +162,7 @@ class TestBinaryTree(unittest.TestCase):
         self.assertEqual(btree.right.right.data, 5)
 
     def test_build_balanced_tree(self):
-        btree = BinaryTree(0)
+        btree = BinarySearchTree(0)
         btree = btree.build_balanced_tree([0, 1, 2, 3, 4, 5], 0, 5)
         self.assertEqual(btree.data, 2)
         self.assertEqual(btree.left.data, 0)
@@ -164,7 +172,7 @@ class TestBinaryTree(unittest.TestCase):
         self.assertEqual(btree.right.right.data, 5)
 
     def test_invert(self):
-        btree = BinaryTree(2)
+        btree = BinarySearchTree(2)
         btree.insert(1)
         btree.insert(3)
         btree.insert(0)
@@ -177,7 +185,7 @@ class TestBinaryTree(unittest.TestCase):
         self.assertEqual(btree.right.left, None)
 
     def test_find_height(self):
-        btree = BinaryTree(2)
+        btree = BinarySearchTree(2)
         self.assertEqual(btree.find_height(), 1)
         btree.insert(3)
         btree.insert(4)
@@ -186,7 +194,7 @@ class TestBinaryTree(unittest.TestCase):
         self.assertEqual(btree.find_height(), 3)
 
     def test_is_balanced(self):
-        btree = BinaryTree(2)
+        btree = BinarySearchTree(2)
         self.assertTrue(btree.is_balanced())
         btree.insert(3)
         self.assertTrue(btree.is_balanced())
