+++ DoublyLinkedList/doubly_linked_list.py
@@ -5,13 +5,20 @@ class Node:
     def __init__(self, data):
         self.data = data
         self.next = None
+        self.prev = None
 
 
 # Linked List Class
-class SinglyLinkedList:
-    def __init__(self):
+class DoublyLinkedList:
+    def __init__(self, initial_data=None):
         self.head = None
+        self.tail = None
         self.size = 0
+        if isinstance(initial_data, int):
+            self.insert_first(initial_data)
+        elif isinstance(initial_data, list):
+            for num in initial_data:
+                self.insert_last(num)
 
     # Empty the linked list
     def clear(self):
@@ -19,8 +26,9 @@ class SinglyLinkedList:
         while trav:
             next = trav.next
             trav.next = None
+            trav.prev = None
             trav = next
-        self.head = trav = None
+        self.head = self.tail = trav = None
         self.size = 0
 
     # Checks if the linked list is empty
@@ -30,8 +38,15 @@ class SinglyLinkedList:
     # Inserts at the begining
     def insert_first(self, new_node):
         new_node = Node(new_node)
+        if self.is_empty():
+            self.head = self.tail = new_node
+            self.size += 1
+            return
+
         new_node.next = self.head
+        new_node.prev = None
         self.head = new_node
+        self.head.next.prev = new_node
         self.size += 1
 
     # Inserts at the end
@@ -39,25 +54,27 @@ class SinglyLinkedList:
         new_node = Node(new_node)
 
         if self.is_empty():
-            self.head = new_node
+            self.head = self.tail = new_node
             self.size += 1
             return
 
-        last = self.head
-        while last.next:
-            last = last.next
-        last.next = new_node
+        new_node.next = None
+        new_node.prev = self.tail
+        self.tail.next = new_node
+        self.tail = new_node
         self.size += 1
 
     # Inserts at a given position
     def insert_at(self, at, new_elem):
         if at < 0 or at > self.size:
-            raise Exception("Overflow: The given at position is invalid")
+            raise Exception('Overflow: The given "at" postion is invalid')
 
         if at == 0:
-            return self.insert_first(new_elem)
+            self.insert_first(new_elem)
+            return
         if at == self.size:
-            return self.insert_last(new_elem)
+            self.insert_last(new_elem)
+            return
 
         index = 1
         new_node = Node(new_elem)
@@ -67,8 +84,11 @@ class SinglyLinkedList:
             curr_node = curr_node.next
             next_node = next_node.next
             index += 1
+
         new_node.next = next_node
+        new_node.prev = curr_node
         curr_node.next = new_node
+        next_node.prev = new_node
         self.size += 1
 
     # Gets the first element of the list
@@ -81,11 +101,7 @@ class SinglyLinkedList:
     def peek_last(self):
         if (self.is_empty()):
             raise Exception("No elements in the list")
-
-        last = self.head
-        while last.next:
-            last = last.next
-        return last.data
+        return self.tail.data
 
     # Removes the first elem
     def remove_first(self):
@@ -102,23 +118,14 @@ class SinglyLinkedList:
         if self.is_empty():
             raise Exception("Nothing to remove")
 
-        if not self.head.next:
-            temp = self.head
-            self.head = None
-            self.size -= 1
-            return temp.data
-
-        curr_node = self.head
-        next_node = self.head.next
-
-        while next_node.next:
-            curr_node = curr_node.next
-            next_node = next_node.next
-
-        temp = next_node
-        next_node = None
-        curr_node.next = None
+        temp = self.tail
+        self.tail = self.tail.prev
+        if self.tail:
+            self.tail.next = None
         self.size -= 1
+
+        if self.is_empty():
+            self.head = None
         return temp.data
 
     # Removes the first ocurrences of a piece of data
@@ -128,6 +135,8 @@ class SinglyLinkedList:
 
         if self.head.data == to_remove:
             return self.remove_first()
+        if self.tail.data == to_remove:
+            return self.remove_last()
 
         curr_node = self.head
         next_node = self.head.next
@@ -135,7 +144,8 @@ class SinglyLinkedList:
         # Traverse until the node is found
         while next_node:
             if next_node.data == to_remove:
-                temp = next_node  # Save next_node to to be able to return
+                temp = next_node
+                next_node.next.prev = curr_node
                 curr_node.next = next_node.next
                 self.size -= 1
                 return temp.data
@@ -158,28 +168,40 @@ class SinglyLinkedList:
     def contains(self, data):
         return self.index_of(data) != -1
 
-    # Reverses the given linked list
+    # Reverse the given linked list
     def reverse(self):
         prev = None
         curr = self.head
         while curr:
-            temp = curr.next
+            prev = curr.prev
+            curr.prev = curr.next
             curr.next = prev
-            prev = curr
-            curr = temp
+            curr = curr.prev
 
         # Set new head only if it was not empty
         if prev != None:
-            self.head = prev
+            self.head = prev.prev
 
-    # Print all elements
-    def print(self):
-        temp = self.head
-        print("[", end='')
-        while temp:
-            if temp.next:
-                print('"' + str(temp.data) + '", ', end='')
-            else:
-                print('"' + str(temp.data) + '"', end='')
-            temp = temp.next
-        print("] (" + str(self.size) + ")")
+    # Represetantion method
+    def __repr__(self):
+        head = self.head
+        arr = [] * self.size
+        while head:
+            arr.append(head.data)
+            head = head.next
+        return "DoublyLinkedList({})".format(arr.__repr__())
+
+    # Convert to string method
+    def __str__(self):
+        head = self.head
+        res = ''
+        while head:
+            res += str(head.data)
+            if head.next:
+                res += '->'
+            head = head.next
+        return res
+
+    # Len of linked list
+    def __len__(self):
+        return self.size
