+++ SinglyLinkedList/__init__.py
