--- Stacks/__init__.py
@@ -1,1 +1,0 @@
-# This file allows Python to recognize .py files in this folder as modules
